[CS] Track parent generic environments for opened pack elements
We currently track these in the StmtChecker to
ensure the correct depth is maintained, but we
didn't track them in the ConstraintSystem. Track
inner generic environments in the trail, and then
make sure we record a new generic environment when
solving the body of a for loop over a pack expansion.